+++ frontend/src/pages/AboutPage/Information.jsx
@@ -1,30 +1,36 @@
 import {Image,Card, CardHeader, CardBody} from "@nextui-org/react";
 import {Spacer} from "@nextui-org/react";
+import {InfoCard} from "./InfoCard";
 
-import {CustomCard} from "./CustomCard";
 
-let Offer = () =>{
+
+let Information = () =>{
     return(
-      <><div class="Aboutmid" style={{ display: "flex", flexDirection: "column" }}>
+        <><div class="Aboutbutton" style={{ display: "flex", flexDirection: "column" }}>
           <center>
-          <div style={{display:"flex",color:"#D97706", fontSize:"10px", fontWeight:"bold",justifyContent:"center"}}>
-        <h >Our Services</h>
+        <div style={{display:"flex",color:"#D97706", fontSize:"10px", fontWeight:"bold",justifyContent:"center"}}>
+        <h >Our Team</h>
         </div>
           <div style={{display:"flex",color:"black", fontSize:"30px", fontWeight:"bold",justifyContent:"center"}}>
-        <h >What we offer</h>
+        <h >Our Awesome Team
+</h>
         </div>
         <div style={{display:"flex",color:"gray",fontSize:"20px",justifyContent:"center"}}>
-        <h>What we offer</h></div>
+        <h>Why are we doing this
+</h></div>
         <div style={{display:"flex",color:"gray",fontSize:"15px",justifyContent:"center"}}>
-        <p>enjoy and protect a peaceful environment for sharing feelings</p></div>
+        <p>We want to be the small drops to form the warm ocean
+
+</p></div>
         </center>
       </div><div class="cards" >
       <div className="flex" style={{flexDirection:"row", justifyContent:"center", margin:"30px"}}>
         
-      <CustomCard />
+      <InfoCard />
     </div>
 
         </div></>
+
     )
 }
-export default Offer;
+export default Information;
